+++ src/pages/administration/students/index.vue
@@ -305,6 +305,9 @@ export default {
 			pagination: (state) =>
 				state.pagination.default || { limit: 10, total: 0 },
 		}),
+		...mapState("federal-states", {
+			federalState: "federalState",
+		}),
 		schoolInternallyManaged() {
 			return !this.school.isExternal;
 		},
@@ -314,15 +317,26 @@ export default {
 			);
 		},
 		filteredColumns() {
-			// filters out edit/consent column if school is external
+			let editedColumns = this.tableColumns;
+			const states = ["NI", "TH"];
+			// filters out edit column if school is external
 			if (this.school.isExternal) {
-				return this.tableColumns.filter(
+				editedColumns = this.tableColumns.filter(
 					//_id field sets the edit column
-					(col) => col.field !== "_id" && col.field !== "consentStatus"
+					(col) => col.field !== "_id"
 				);
 			}
 
-			return this.tableColumns;
+			// filters out the consent column if the school is from any of the states and is not external
+			if (states.some((state) => this.federalState.abbreviation == state)) {
+				if (!this.school.isExternal) {
+					editedColumns = editedColumns.filter(
+						(col) => col.field !== "consentStatus"
+					);
+				}
+			}
+
+			return editedColumns;
 		},
 	},
 	watch: {
@@ -350,7 +364,9 @@ export default {
 	},
 	created(ctx) {
 		this.find();
-		this.getFederalStateData();
+		if (this.school.federalState) {
+			this.getFederalStateData();
+		}
 	},
 	methods: {
 		find() {